pyproject.toml
[project]
name = "sudo-air"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "influxdb-client>=1.49.0",
    "loguru>=0.7.0",
    "paho-mqtt>=2.1.0",
    "typer>=0.21.1",
]
